requirements.txt
beautifulsoup4==4.5.3
Django==1.10.6
django-filter==1.0.1
django-model-utils==2.6.1
djangorestframework==3.6.2
docopt==0.6.2
factory-boy==2.8.1
Faker==0.7.10
hues==0.2.2
Markdown==2.6.8
munch==2.1.1
progressbar2==3.16.0
psycopg2==2.7.1
python-dateutil==2.6.0
python-utils==2.0.1
pytz==2016.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
